# fix: avoid recreating existing, valid symlinks