@zkochan 2024-08-19
I am afraid to merge this as it can make install slower in other scenarios. Like with clean install.