@stevenpetryk 2024-08-20
> I am afraid to merge this as it can make install slower in other scenarios. Like with clean install.

I'll do some benchmarks on Discord's repo, with and without our (equivalent) patch.