@stevenpetryk 2024-08-20
Alright, I did the following:

- Moved virtual store to /tmp (to make rm -rf'ing node_modules easier)
- Used the same exact pnpm version with and without the patch
- Gave each version 5 warmup `pnpm install`s

Mysteriously, no matter how many times I try, with this patch, a fresh install is marginally _faster_. It makes me wonder if something else is going on (maybe pnpm tries to double-write some dirs?)

Without patch:

```
~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.94s user 2.26s system 190% cpu 3.251 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.03s user 2.42s system 198% cpu 3.254 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.03s user 2.50s system 202% cpu 3.216 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.89s user 2.39s system 201% cpu 3.109 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.93s user 2.34s system 199% cpu 3.134 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.03s user 2.57s system 201% cpu 3.269 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.01s user 2.65s system 200% cpu 3.317 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.02s user 2.32s system 192% cpu 3.294 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.98s user 2.39s system 193% cpu 3.292 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  4.01s user 2.39s system 192% cpu 3.331 total
```

With patch:

```
~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.82s user 1.65s system 181% cpu 3.006 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.88s user 1.65s system 183% cpu 3.012 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.76s user 1.58s system 184% cpu 2.894 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.79s user 1.65s system 178% cpu 3.045 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.87s user 1.65s system 185% cpu 2.969 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.77s user 1.57s system 186% cpu 2.862 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.86s user 1.66s system 184% cpu 2.996 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.92s user 1.77s system 182% cpu 3.114 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.96s user 1.82s system 185% cpu 3.113 total

~/src/discord main* ⇣
❯ rm -rf **/node_modules && time ./result/bin/pnpm install --silent
./result/bin/pnpm install --silent  3.96s user 1.66s system 183% cpu 3.053 total
```