@zkochan 2024-08-24
For me it shows worth results, when testing the change on https://github.com/pnpm/pnpm.io/blob/main/benchmarks/fixtures/alotta-files/package.json