@zkochan 2024-08-24
Will this work? https://github.com/pnpm/symlink-dir/pull/54